Refactor sensor data generation to return integer values for temperature, humidity, and luminosity

+++ SimuladorArduino.java
@@ -1,9 +1,9 @@
 public class SimuladorArduino {
    
     public static String [] leitura(){
-        String temperatura = String.valueOf(Math.random() * 100);
-        String umidade = String.valueOf(Math.random() * 100);
-        String luminosidade = String.valueOf(Math.random() * 100);
+        String temperatura = String.valueOf((int) (Math.random() * 100));
+        String umidade = String.valueOf((int) (Math.random() * 100));
+        String luminosidade = String.valueOf((int) (Math.random() * 100));
         return new String[] {temperatura, umidade, luminosidade};
     }
